refactor(vocab-match): split VocabDefinitionMatch.tsx 1005L → orchestrator + 5 modules (Related to #1846)
TDD-first: 26 characterization tests written before extraction, tautology verified.

Extracted:
- vocabDefinitionMatchLogic.ts (132L): shuffle, getDragDropAttemptFeedback, getDragDropAttemptTone,
  buildMCQOptions, selectRetryIndices, mergePersistedProgress — all pure, zero React
- VocabDefinitionMatchSummary.tsx (155L): SummaryScreen — score card + per-question results
- VocabDefinitionMatchStageStatus.tsx (75L): StageStatus — two-step progress indicator
- VocabDefinitionMatchMCQ.tsx (106L): MultipleChoiceMode — MCQ interaction with #1101 distractor fix
- VocabDefinitionMatchDragDrop.tsx (346L): DragDropMode — drag/touch interaction with #1101/#1102 fixes

VocabDefinitionMatch.tsx: 1005L → 282L thin orchestrator, no behavior change.
Build: zero TS errors. Tests: 26/26 green.
